+++ src/main/kotlin/org/olafneumann/regex/generator/ui/parts/LanguageAccordionCard.kt
@@ -1,4 +1,4 @@
-package org.olafneumann.regex.generator.ui.html
+package org.olafneumann.regex.generator.ui.parts
 
 import kotlinx.browser.document
 import kotlinx.html.ButtonType
@@ -14,6 +14,7 @@ import org.olafneumann.regex.generator.js.jQuery
 import org.olafneumann.regex.generator.output.CodeGenerator
 import org.olafneumann.regex.generator.output.CodeGenerator.GeneratedSnippet
 import org.olafneumann.regex.generator.settings.ApplicationSettings
+import org.olafneumann.regex.generator.ui.utils.HtmlHelper
 import org.w3c.dom.HTMLElement
 import org.w3c.dom.Node
 
@@ -45,8 +46,8 @@ internal class LanguageAccordionCard(
                 }
             }
         })
-        codeElement = document.getElementById(codeElementId) as HTMLElement
-        bodyElement = document.getElementById(bodyElementId) as HTMLElement
+        codeElement = HtmlHelper.getElementById(codeElementId)
+        bodyElement = HtmlHelper.getElementById(bodyElementId)
 
         jQuery(bodyElement).on("shown.bs.collapse") { shown = true }
         jQuery(bodyElement).on("hidden.bs.collapse") { shown = false }
