+++ src/main/kotlin/org/olafneumann/regex/generator/ui/utils/TimerController.kt
@@ -1,4 +1,4 @@
-package org.olafneumann.regex.generator.ui.html
+package org.olafneumann.regex.generator.ui.utils
 
 import kotlinx.browser.window
